pyproject.toml
[tool.poetry]
name = "NewHorizons"
version = "0.1.0"
description = ""
authors = ["Edd Salkield <edd@salkield.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
youtube-dl = "^2021.3.25"
fastapi = "^0.63.0"
httpx = "^0.17.1"
beautifulsoup4 = "^4.9.3"
"calmjs.parse" = "^1.2.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
uvicorn = "^0.13.4"

[tool.poetry.scripts]
test-server = 'uvicorn newhorizons.main:app'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

